+++ shared/src/commonMain/kotlin/domain/BuildAuthenticationConsentPageFromAuthenticationRequest.kt
@@ -5,11 +5,14 @@ import at.asitplus.catching
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.wallet.app.common.PresentationService
+import at.asitplus.wallet.app.common.relyingParty.WrpValidationResult
+import at.asitplus.wallet.app.common.relyingParty.WrpValidator
 import io.github.aakira.napier.Napier
 import ui.navigation.routes.AuthenticationViewRoute
 
 class BuildAuthenticationConsentPageFromAuthenticationRequest(
     val presentationService: PresentationService,
+    val wrpValidator: WrpValidator
 ) {
     suspend operator fun invoke(
         request: RequestParametersFrom<AuthenticationRequestParameters>,
@@ -17,12 +20,16 @@ class BuildAuthenticationConsentPageFromAuthenticationRequest(
         val preparationState = presentationService.startAuthorizationResponsePreparation(request)
             .onFailure { Napier.e("Failure", it) }
             .getOrThrow()
-
+        val wrpRequestValidationResult = wrpValidator.validate(preparationState.request).getOrElse {
+            Napier.w("WRP Validation failed, continuing without registration certificate.", throwable = it)
+            null
+        }
         AuthenticationViewRoute(
             authenticationRequest = preparationState.request,
             authorizationResponsePreparationState = preparationState,
             recipientLocation = preparationState.request.parameters.clientId ?: "",
             isCrossDeviceFlow = false,
+            wrpValidationResult = wrpRequestValidationResult
         )
     }
 }