+++ shared/src/commonMain/kotlin/ui/presentation/CredentialSetQueryOptionSelectionCardCredentialQueryContent.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.additional_text_other_untranslated_claims
@@ -18,6 +19,7 @@ fun ColumnScope.CredentialSetQueryOptionSelectionCardCredentialQueryContent(
     credentialSchemeLocalized: String,
     credentialAttributesLocalized: Pair<List<String>, Int>?,
     credentialRepresentationLocalized: String? = null,
+    credentialAllowedAttributes:  Map<String, Boolean>? = null,
 ) {
     if (credentialRepresentationLocalized != null) {
         LabeledText(
@@ -32,7 +34,12 @@ fun ColumnScope.CredentialSetQueryOptionSelectionCardCredentialQueryContent(
             modifier = Modifier.padding(start = 8.dp)
         ) {
             attributeNames.forEachIndexed { index, it ->
-                Text(it)
+                val allowed = credentialAllowedAttributes?.get(it) ?: true
+                val color = when(allowed) {
+                    false -> Color.Red
+                    else -> Color.Unspecified
+                }
+                Text(text = it, color = color)
             }
             if (otherClaimReferences > 0) {
                 Text("+$otherClaimReferences " + stringResource(Res.string.additional_text_other_untranslated_claims))